+++ precompiles/outposts/osmosis/tx_test.go
@@ -4,9 +4,11 @@ import (
 	"fmt"
 	"math/big"
 
-	sdktypes "github.com/cosmos/cosmos-sdk/types"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/vm"
+
+	sdktypes "github.com/cosmos/cosmos-sdk/types"
+
 	"github.com/evmos/evmos/v15/precompiles/authorization"
 	cmn "github.com/evmos/evmos/v15/precompiles/common"
 	"github.com/evmos/evmos/v15/precompiles/ics20"
@@ -48,7 +50,8 @@ func (s *PrecompileTestSuite) TestSwap() {
 			},
 			expError:    true,
 			errContains: fmt.Sprintf(cmn.ErrInvalidNumberOfArgs, 7, 0),
-		}, {
+		},
+		{
 			name:   "fail - origin different from sender",
 			sender: senderAddress,
 			origin: senderAddress,
@@ -65,7 +68,8 @@ func (s *PrecompileTestSuite) TestSwap() {
 			},
 			expError:    true,
 			errContains: fmt.Sprintf(ics20.ErrDifferentOriginFromSender, senderAddress, randomAddress),
-		}, {
+		},
+		{
 			name:   "fail - missing input token denom",
 			sender: senderAddress,
 			origin: senderAddress,
@@ -85,7 +89,8 @@ func (s *PrecompileTestSuite) TestSwap() {
 			},
 			expError:    true,
 			errContains: fmt.Sprintf("token '%s' not registered", randomAddress),
-		}, {
+		},
+		{
 			name:   "fail - missing output token denom",
 			sender: senderAddress,
 			origin: senderAddress,
@@ -105,7 +110,8 @@ func (s *PrecompileTestSuite) TestSwap() {
 			},
 			expError:    true,
 			errContains: fmt.Sprintf("token '%s' not registered", randomAddress),
-		}, {
+		},
+		{
 			name:   "fail - osmo token pair not registered (with osmo hardcoded address)",
 			sender: senderAddress,
 			origin: senderAddress,
@@ -125,19 +131,20 @@ func (s *PrecompileTestSuite) TestSwap() {
 			},
 			expError:    true,
 			errContains: fmt.Sprintf("token '%s' not registered", common.HexToAddress("0x1D54EcB8583Ca25895c512A8308389fFD581F9c9")),
-		}, {
-			name:   "fail - osmo token pair registered with another channelID",
+		},
+		{
+			name:   "fail - osmo token pair registered with another ChannelID",
 			sender: senderAddress,
 			origin: senderAddress,
 			malleate: func() []interface{} {
 				evmosTokenPair, err := testutils.RegisterEvmosERC20Coins(*s.unitNetwork, sender)
 				s.Require().NoError(err, "expected no error during evmos erc20 registration")
 
-				osmoIbcDenomTrace := utils.ComputeIBCDenomTrace(portID, channelID, osmosis.OsmosisDenom)
+				osmoIbcDenomTrace := utils.ComputeIBCDenomTrace(PortID, ChannelID, osmosis.OsmosisDenom)
 				_, err = testutils.RegisterIBCERC20Coins(*s.unitNetwork, sender, osmoIbcDenomTrace)
 				s.Require().NoError(err, "expected no error during ibc erc20 registration")
 
-				wrongOsmoIbcDenomTrace := utils.ComputeIBCDenomTrace(portID, "channel-1", osmosis.OsmosisDenom)
+				wrongOsmoIbcDenomTrace := utils.ComputeIBCDenomTrace(PortID, "channel-1", osmosis.OsmosisDenom)
 				wrongOsmoTokenPair, err := testutils.RegisterIBCERC20Coins(*s.unitNetwork, sender, wrongOsmoIbcDenomTrace)
 				s.Require().NoError(err, "expected no error during ibc erc20 registration")
 
@@ -153,8 +160,9 @@ func (s *PrecompileTestSuite) TestSwap() {
 			},
 			expError: true,
 			// Probably there is a better way than hardcoding the expected string
-			errContains: fmt.Sprintf(osmosis.ErrInputTokenNotSupported, []string{"aevmos", "ibc/ED07A3391A112B175915CD8FAF43A2DA8E4790EDE12566649D0C2F97716B8518"}),
-		}, {
+			errContains: fmt.Sprintf(osmosis.ErrDenomNotSupported, []string{utils.BaseDenom, "ibc/ED07A3391A112B175915CD8FAF43A2DA8E4790EDE12566649D0C2F97716B8518"}),
+		},
+		{
 			name:   "fail - input equal to denom",
 			sender: senderAddress,
 			origin: senderAddress,
@@ -173,8 +181,9 @@ func (s *PrecompileTestSuite) TestSwap() {
 				}
 			},
 			expError:    true,
-			errContains: fmt.Sprintf(osmosis.ErrInputEqualOutput, "aevmos"),
-		}, {
+			errContains: fmt.Sprintf(osmosis.ErrInputEqualOutput, utils.BaseDenom),
+		},
+		{
 			name:   "fail - invalid input",
 			sender: senderAddress,
 			origin: senderAddress,
@@ -182,7 +191,7 @@ func (s *PrecompileTestSuite) TestSwap() {
 				evmosTokenPair, err := testutils.RegisterEvmosERC20Coins(*s.unitNetwork, sender)
 				s.Require().NoError(err, "expected no error during evmos erc20 registration")
 
-				wrongIbcDenomTrace := utils.ComputeIBCDenomTrace(portID, channelID, "wrong")
+				wrongIbcDenomTrace := utils.ComputeIBCDenomTrace(PortID, ChannelID, "wrong")
 				wrongTokenPair, err := testutils.RegisterIBCERC20Coins(*s.unitNetwork, sender, wrongIbcDenomTrace)
 				s.Require().NoError(err, "expected no error during ibc erc20 registration")
 
@@ -198,8 +207,37 @@ func (s *PrecompileTestSuite) TestSwap() {
 			},
 			expError: true,
 			// Probably there is a better way than hardcoding the expected string
-			errContains: fmt.Sprintf(osmosis.ErrInputTokenNotSupported, []string{"aevmos", "ibc/ED07A3391A112B175915CD8FAF43A2DA8E4790EDE12566649D0C2F97716B8518"}),
-		}, {
+			errContains: fmt.Sprintf(osmosis.ErrDenomNotSupported, []string{utils.BaseDenom, "ibc/ED07A3391A112B175915CD8FAF43A2DA8E4790EDE12566649D0C2F97716B8518"}),
+		},
+		// All tests below requires the ibcSetup equal to true because run the query GetChannel
+		// that fails if the IBC channel is not open.
+		{
+			name:   "fail - channel not open",
+			sender: senderAddress,
+			origin: senderAddress,
+			malleate: func() []interface{} {
+				evmosTokenPair, err := testutils.RegisterEvmosERC20Coins(*s.unitNetwork, sender)
+				s.Require().NoError(err, "expected no error during evmos erc20 registration")
+
+				osmoIbcDenomTrace := utils.ComputeIBCDenomTrace(PortID, ChannelID, osmosis.OsmosisDenom)
+				osmoTokenPair, err := testutils.RegisterIBCERC20Coins(*s.unitNetwork, sender, osmoIbcDenomTrace)
+				s.Require().NoError(err, "expected no error during ibc erc20 registration")
+
+				return []interface{}{
+					senderAddress,
+					osmoTokenPair.GetERC20Contract(),
+					evmosTokenPair.GetERC20Contract(),
+					transferAmount,
+					slippagePercentage,
+					windowSeconds,
+					receiver,
+				}
+			},
+			expError:    true,
+			ibcSetup:    false,
+			errContains: fmt.Sprintf("port ID (%s) channel ID (%s)", PortID, ChannelID),
+		},
+		{
 			name:   "fail - receiver is not a valid bech32",
 			sender: senderAddress,
 			origin: senderAddress,
@@ -207,7 +245,7 @@ func (s *PrecompileTestSuite) TestSwap() {
 				evmosTokenPair, err := testutils.RegisterEvmosERC20Coins(*s.unitNetwork, sender)
 				s.Require().NoError(err, "expected no error during evmos erc20 registration")
 
-				osmoIbcDenomTrace := utils.ComputeIBCDenomTrace(portID, channelID, osmosis.OsmosisDenom)
+				osmoIbcDenomTrace := utils.ComputeIBCDenomTrace(PortID, ChannelID, osmosis.OsmosisDenom)
 				osmoTokenPair, err := testutils.RegisterIBCERC20Coins(*s.unitNetwork, sender, osmoIbcDenomTrace)
 				s.Require().NoError(err, "expected no error during ibc erc20 registration")
 
@@ -222,8 +260,10 @@ func (s *PrecompileTestSuite) TestSwap() {
 				}
 			},
 			expError:    true,
+			ibcSetup:    true,
 			errContains: fmt.Sprintf(osmosis.ErrReceiverAddress, "not a valid evmos address"),
-		}, {
+		},
+		{
 			//  THIS PANICS INSIDE CheckAuthzExists
 			name:   "fail - origin different from address caller",
 			sender: senderAddress,
@@ -232,7 +272,7 @@ func (s *PrecompileTestSuite) TestSwap() {
 				evmosTokenPair, err := testutils.RegisterEvmosERC20Coins(*s.unitNetwork, sender)
 				s.Require().NoError(err, "expected no error during evmos erc20 registration")
 
-				osmoIbcDenomTrace := utils.ComputeIBCDenomTrace(portID, channelID, osmosis.OsmosisDenom)
+				osmoIbcDenomTrace := utils.ComputeIBCDenomTrace(PortID, ChannelID, osmosis.OsmosisDenom)
 				osmoTokenPair, err := testutils.RegisterIBCERC20Coins(*s.unitNetwork, sender, osmoIbcDenomTrace)
 				s.Require().NoError(err, "expected no error during ibc erc20 registration")
 
@@ -247,8 +287,10 @@ func (s *PrecompileTestSuite) TestSwap() {
 				}
 			},
 			expError:    true,
+			ibcSetup:    true,
 			errContains: fmt.Sprintf(authorization.ErrAuthzDoesNotExistOrExpired, senderAddress, s.keyring.GetAddr(1)),
-		}, {
+		},
+		{
 			name:   "fail - ibc channel not open",
 			sender: senderAddress,
 			origin: senderAddress,
@@ -256,7 +298,7 @@ func (s *PrecompileTestSuite) TestSwap() {
 				evmosTokenPair, err := testutils.RegisterEvmosERC20Coins(*s.unitNetwork, sender)
 				s.Require().NoError(err, "expected no error during evmos erc20 registration")
 
-				osmoIbcDenomTrace := utils.ComputeIBCDenomTrace(portID, channelID, osmosis.OsmosisDenom)
+				osmoIbcDenomTrace := utils.ComputeIBCDenomTrace(PortID, ChannelID, osmosis.OsmosisDenom)
 				osmoTokenPair, err := testutils.RegisterIBCERC20Coins(*s.unitNetwork, sender, osmoIbcDenomTrace)
 				s.Require().NoError(err, "expected no error during ibc erc20 registration")
 
@@ -271,16 +313,17 @@ func (s *PrecompileTestSuite) TestSwap() {
 				}
 			},
 			expError:    true,
-			errContains: fmt.Sprintf("port ID (%s) channel ID (%s)", portID, channelID),
-		}, {
-			name:   "pass - correct swap",
+			errContains: fmt.Sprintf("port ID (%s) channel ID (%s)", PortID, ChannelID),
+		},
+		{
+			name:   "pass - correct swap output ibc evmos",
 			sender: senderAddress,
 			origin: senderAddress,
 			malleate: func() []interface{} {
 				evmosTokenPair, err := testutils.RegisterEvmosERC20Coins(*s.unitNetwork, sender)
 				s.Require().NoError(err, "expected no error during evmos erc20 registration")
 
-				osmoIbcDenomTrace := utils.ComputeIBCDenomTrace(portID, channelID, osmosis.OsmosisDenom)
+				osmoIbcDenomTrace := utils.ComputeIBCDenomTrace(PortID, ChannelID, osmosis.OsmosisDenom)
 				osmoTokenPair, err := testutils.RegisterIBCERC20Coins(*s.unitNetwork, sender, osmoIbcDenomTrace)
 				s.Require().NoError(err, "expected no error during ibc erc20 registration")
 
@@ -297,12 +340,36 @@ func (s *PrecompileTestSuite) TestSwap() {
 			expError: false,
 			ibcSetup: true,
 		},
+		{
+			name:   "pass - correct swap output osmo",
+			sender: senderAddress,
+			origin: senderAddress,
+			malleate: func() []interface{} {
+				evmosTokenPair, err := testutils.RegisterEvmosERC20Coins(*s.unitNetwork, sender)
+				s.Require().NoError(err, "expected no error during evmos erc20 registration")
+
+				osmoIbcDenomTrace := utils.ComputeIBCDenomTrace(PortID, ChannelID, osmosis.OsmosisDenom)
+				osmoTokenPair, err := testutils.RegisterIBCERC20Coins(*s.unitNetwork, sender, osmoIbcDenomTrace)
+				s.Require().NoError(err, "expected no error during ibc erc20 registration")
+
+				return []interface{}{
+					senderAddress,
+					evmosTokenPair.GetERC20Contract(),
+					osmoTokenPair.GetERC20Contract(),
+					transferAmount,
+					slippagePercentage,
+					windowSeconds,
+					receiver,
+				}
+			},
+			expError: false,
+			ibcSetup: true,
+		},
 	}
 
 	for _, tc := range testCases {
 		s.Run(tc.name, func() {
 			s.SetupTest()
-
 			contract := vm.NewContract(vm.AccountRef(tc.sender), s.precompile, big.NewInt(0), gas)
 
 			stateDB := s.unitNetwork.GetStateDB()